server.py was really getting big, split views into modules
