+++ app/models.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python2
 # -*-coding:UTF-8 -*
 
-import ConfigParser
 from datetime import datetime
 
 from flask import Flask, current_app
@@ -11,26 +10,7 @@ from flask.ext.login import UserMixin, AnonymousUserMixin
 from flask.ext.scrypt import generate_random_salt, generate_password_hash, check_password_hash
 from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 
-
-# Parsing Conf File #
-configfile = 'config/config.cfg'
-cfg = ConfigParser.ConfigParser()
-cfg.read(configfile)
-
-# Flask app, bootstrap etc. #
-app = Flask(__name__)
-app.config['SECRET_KEY'] = cfg.get("Secret", "wtf_secret_key")
-
-db_type = cfg.get('Database', 'type')
-db_user = cfg.get('Database', 'user')
-db_pwd = cfg.get('Database', 'password')
-db_hostname = cfg.get('Database', 'hostname')
-db_name = cfg.get('Database', 'db_name')
-
-app.config['SQLALCHEMY_DATABASE_URI'] = db_type + '://' + db_user + ':' + db_pwd + '@' + db_hostname + '/' + db_name
-app.config['SQLALCHEMY_COMMIT_ON_TEARDOWN'] = True
-
-db = SQLAlchemy(app)
+db = SQLAlchemy()
 
 
 class Permission(object):
