+++ app/templates/auth/login.html
@@ -16,8 +16,8 @@
             <div class="panel-body">
                 {{ wtf.quick_form(form) }}
                 </br>
-                <h6>Forgot your password? <a href="{{ url_for('password_reset_request') }}">Click here to reset it</a>.</h6>
-                <h6>New user? <a href="{{ url_for('register') }}">Click here to register</a>.</h6>
+                <h6>Forgot your password? <a href="{{ url_for('user.password_reset_request') }}">Click here to reset it</a>.</h6>
+                <h6>New user? <a href="{{ url_for('user.register') }}">Click here to register</a>.</h6>
             </div>
         </div>
     </div>
