+++ app/templates/auth/unconfirmed.html
@@ -11,6 +11,6 @@ Check your inbox, you should have received an email with a confirmation link.
 </p>
 <p>
 Need another confirmation email?
-<a href="{{ url_for('resend_confirmation') }}">Click here</a>
+<a href="{{ url_for('user.resend_confirmation') }}">Click here</a>
 </p>
 {% endblock %}
