+++ app/templates/base.html
@@ -13,7 +13,7 @@
         <ul class="nav navbar-nav navbar-right">
             {% if current_user.is_authenticated %}
             <li>
-            <a href='{{ url_for('user') }}' class="nav">
+            <a href='{{ url_for('user.user') }}' class="nav">
                 <span class="glyphicon glyphicon-bullhorn"></span> Notifications
             </a>
             </li>
@@ -27,14 +27,14 @@
             <li class='dropdown'>
             <a href="#" class="dropdown-toggle" data-toggle="dropdown"><span class="glyphicon glyphicon-user"></span> Account {{ current_user.email }}<b class="caret"></b></a>
             <ul class="dropdown-menu">
-                <li><a href='{{ url_for('change_email_request') }}'><span class="glyphicon glyphicon-envelope" aria-hidden="true"></span> Change Email</a></li>
-                <li><a href='{{ url_for('change_pgp') }}'><span class="glyphicon glyphicon-qrcode" aria-hidden="true"></span> Change PGP</a></li>
-                <li><a href='{{ url_for('change_password') }}'><span class="glyphicon glyphicon-lock" aria-hidden="true"></span> Change Password</a></li>
+                <li><a href='{{ url_for('user.change_email_request') }}'><span class="glyphicon glyphicon-envelope" aria-hidden="true"></span> Change Email</a></li>
+                <li><a href='{{ url_for('user.change_pgp') }}'><span class="glyphicon glyphicon-qrcode" aria-hidden="true"></span> Change PGP</a></li>
+                <li><a href='{{ url_for('user.change_password') }}'><span class="glyphicon glyphicon-lock" aria-hidden="true"></span> Change Password</a></li>
                 <li class="divider"></li>
-                <li><a href="{{ url_for('logout') }}"><span class="glyphicon glyphicon-log-out" aria-hidden="true"></span> Logout</a></li>
+                <li><a href="{{ url_for('user.logout') }}"><span class="glyphicon glyphicon-log-out" aria-hidden="true"></span> Logout</a></li>
                 {% else %}
-                <li><a href="{{ url_for('register') }}"><span class="glyphicon glyphicon-plus" aria-hidden="true"></span> Register</a></li>
-                <li><a href="{{ url_for('login') }}"><span class="glyphicon glyphicon-log-in" aria-hidden="true"></span> Login</a></li>
+                <li><a href="{{ url_for('user.register') }}"><span class="glyphicon glyphicon-plus" aria-hidden="true"></span> Register</a></li>
+                <li><a href="{{ url_for('user.login') }}"><span class="glyphicon glyphicon-log-in" aria-hidden="true"></span> Login</a></li>
                 {% endif %}
             </ul>
         </div>
