+++ app/templates/emails/change_email.html
@@ -1,7 +1,7 @@
 <p>Dear {{ user.name }},</p>
-<p>To confirm your new email address <a href="{{ url_for('change_email', token=token, _external=True) }}">click here</a>.</p>
+<p>To confirm your new email address <a href="{{ url_for('user.change_email', token=token, _external=True) }}">click here</a>.</p>
 <p>Alternatively, you can paste the following link in your browser's address bar:</p>
-<p>{{ url_for('change_email', token=token, _external=True) }}</p>
+<p>{{ url_for('user.change_email', token=token, _external=True) }}</p>
 <p>Sincerely,</p>
 <p>The CVE-Portal Team</p>
 <p><small>Note: replies to this email address are not monitored.</small></p>
