+++ app/templates/emails/change_email.txt
@@ -1,7 +1,7 @@
 Dear {{ user.name }},
 To confirm your new email address click on the following link:
 
-{{ url_for('change_email', token=token, _external=True) }}
+{{ url_for('user.change_email', token=token, _external=True) }}
 
 Sincerely,
