+++ app/templates/emails/confirm.html
@@ -2,7 +2,7 @@
 <p>Welcome to <b>CVE-Portal</b>!</p>
 <p>To confirm your account please <a href="{{ url_for('confirm', token=token, _external=True) }}">click here</a>.</p>
 <p>Alternatively, you can paste the following link in your browser's address bar:</p>
-<p>{{ url_for('confirm', token=token, _external=True) }}</p>
+<p>{{ url_for('user.confirm', token=token, _external=True) }}</p>
 <p>Sincerely,</p>
 <p>The CVE-Portal Team</p>
 <p><small>Note: replies to this email address are not monitored.</small></p>
