+++ app/templates/emails/confirm.txt
@@ -3,7 +3,7 @@ Dear {{ user.name }},
 Welcome to CVE-Portal!
 To confirm your account please click on the following link:
 
-{{ url_for('confirm', token=token, _external=True) }}
+{{ url_for('user.confirm', token=token, _external=True) }}
 
 Sincerely,
