+++ app/templates/emails/password_reset.html
@@ -1,7 +1,7 @@
 <p>Dear {{ user.username }},</p>
-<p>To reset your password <a href="{{ url_for('password_reset', token=token, _external=True) }}">click here</a>.</p>
+<p>To reset your password <a href="{{ url_for('user.password_reset', token=token, _external=True) }}">click here</a>.</p>
 <p>Alternatively, you can paste the following link in your browser's address bar:</p>
-<p>{{ url_for('password_reset', token=token, _external=True) }}</p>
+<p>{{ url_for('user.password_reset', token=token, _external=True) }}</p>
 <p>If you have not requested a password reset simply ignore this message.</p>
 <p>Sincerely,</p>
 <p>The CVE-Portal Team</p>
