+++ app/templates/emails/password_reset.txt
@@ -1,7 +1,7 @@
 Dear {{ user.name }},
 To reset your password click on the following link:
 
-{{ url_for('password_reset', token=token, _external=True) }}
+{{ url_for('user.password_reset', token=token, _external=True) }}
 
 If you have not requested a password reset simply ignore this message.
